+++ src/main/java/network/crypta/node/OpennetPeerNode.java
@@ -3,8 +3,8 @@ package network.crypta.node;
 import network.crypta.io.comm.PeerParseException;
 import network.crypta.io.comm.ReferenceSignatureVerificationException;
 import network.crypta.node.OpennetManager.LinkLengthClass;
-import network.crypta.node.updater.NodeUpdateManager;
-import network.crypta.node.updater.UpdateOverMandatoryManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.UpdateOverMandatoryManager;
 import network.crypta.support.SimpleFieldSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
