+++ src/main/java/network/crypta/runtime/updater/NodeUpdaterParams.java
@@ -1,4 +1,4 @@
-package network.crypta.node.updater;
+package network.crypta.runtime.updater;
 
 import network.crypta.keys.FreenetURI;
