+++ src/main/java/network/crypta/runtime/updater/UpdateOverMandatoryManager.java
@@ -1,4 +1,4 @@
-package network.crypta.node.updater;
+package network.crypta.runtime.updater;
 
 import java.io.DataInputStream;
 import java.io.EOFException;
