+++ src/test/java/network/crypta/node/AnnouncerTest.java
@@ -5,10 +5,10 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
-import network.crypta.node.updater.NodeUpdateManager;
-import network.crypta.node.updater.UpdateOverMandatoryManager;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.UpdateOverMandatoryManager;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.Ticker;
