+++ src/test/java/network/crypta/node/NodeControlMessageHandlerTest.java
@@ -9,8 +9,8 @@ import network.crypta.io.comm.Peer;
 import network.crypta.node.subsystem.NodeMessagingSubsystem;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
-import network.crypta.node.updater.NodeUpdateManager;
-import network.crypta.node.updater.UpdateOverMandatoryManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.UpdateOverMandatoryManager;
 import network.crypta.support.Fields;
 import network.crypta.support.ShortBuffer;
 import network.crypta.support.Ticker;
