+++ src/test/java/network/crypta/node/NodeDispatcherTest.java
@@ -10,8 +10,8 @@ import network.crypta.io.comm.PeerContext;
 import network.crypta.keys.NodeCHK;
 import network.crypta.node.NodeDispatcher.NodeDispatcherCallback;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
-import network.crypta.node.updater.NodeUpdateManager;
-import network.crypta.node.updater.UpdateOverMandatoryManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.UpdateOverMandatoryManager;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.ShortBuffer;
 import network.crypta.support.Ticker;
