+++ src/test/java/network/crypta/node/OpennetPeerNodeTest.java
@@ -10,8 +10,8 @@ import network.crypta.io.comm.MessageCore;
 import network.crypta.io.comm.UdpSocketHandler;
 import network.crypta.node.OpennetManager.ConnectionType;
 import network.crypta.node.OpennetManager.LinkLengthClass;
-import network.crypta.node.updater.NodeUpdateManager;
-import network.crypta.node.updater.UpdateOverMandatoryManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.UpdateOverMandatoryManager;
 import network.crypta.support.BooleanLastTrueTracker;
 import network.crypta.support.SimpleFieldSet;
 import org.junit.jupiter.api.BeforeEach;
