+++ src/test/java/network/crypta/node/PeerAlertCoordinatorTest.java
@@ -1,10 +1,10 @@
 package network.crypta.node;
 
 import java.lang.reflect.Field;
-import network.crypta.node.updater.NodeUpdateManager;
 import network.crypta.runtime.alerts.PeerManagerUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
