+++ src/test/java/network/crypta/node/subsystem/NodeServicesSubsystemTest.java
@@ -10,7 +10,6 @@ import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.NodeInitException;
 import network.crypta.node.diagnostics.DefaultNodeDiagnostics;
-import network.crypta.node.updater.NodeUpdateManager;
 import network.crypta.runtime.alerts.JVMVersionAlert;
 import network.crypta.runtime.alerts.MeaningfulNodeNameUserAlert;
 import network.crypta.runtime.alerts.NotEnoughNiceLevelsUserAlert;
@@ -19,6 +18,7 @@ import network.crypta.runtime.alerts.SimpleUserAlert;
 import network.crypta.runtime.alerts.TimeSkewDetectedUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.Ticker;
