+++ src/test/java/network/crypta/runtime/admin/LegacyConnectionsSupportPortTest.java
@@ -10,8 +10,8 @@ import network.crypta.node.NodeFile;
 import network.crypta.node.ProgramDirectory;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeServicesSubsystem;
-import network.crypta.node.updater.NodeUpdateManager;
 import network.crypta.runtime.spi.ConnectionsInstallerSnapshot;
+import network.crypta.runtime.updater.NodeUpdateManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
