+++ src/test/java/network/crypta/runtime/admin/LegacyWelcomeActionPortTest.java
@@ -5,10 +5,10 @@ import network.crypta.config.PersistentConfig;
 import network.crypta.config.SubConfig;
 import network.crypta.node.Node;
 import network.crypta.node.NodeStarter;
-import network.crypta.node.updater.NodeUpdateManager;
 import network.crypta.runtime.alerts.UpgradeConnectionSpeedUserAlert;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
+import network.crypta.runtime.updater.NodeUpdateManager;
 import network.crypta.support.Ticker;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
