+++ src/test/java/network/crypta/runtime/core/LegacyCoreUpdateActionPortTest.java
@@ -3,8 +3,8 @@ package network.crypta.runtime.core;
 import java.io.File;
 import java.nio.file.Path;
 import network.crypta.node.Node;
-import network.crypta.node.updater.CoreUpdater;
-import network.crypta.node.updater.NodeUpdateManager;
+import network.crypta.runtime.updater.CoreUpdater;
+import network.crypta.runtime.updater.NodeUpdateManager;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
