+++ src/test/java/network/crypta/runtime/updater/CoreUpdaterTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.updater;
+package network.crypta.runtime.updater;
 
 import org.junit.jupiter.api.Test;
